Fix Unsupported Operation error in Tally widget
This fixes an exception printing in the console when hot-reloading the
application in the IDE:
```
════════ Exception caught by widgets library ═══════════════════════════════════════════════════════
The following UnsupportedError was thrown building LinearProgressIndicator(100.0%, dirty, dependencies: [TickerMode, Directionality], state: _LinearProgressIndicatorState#50427(ticker inactive and muted)):
Unsupported operation: Infinity or NaN toInt

User-created ancestor of the error-causing widget was:
  Container file:///Users/pieper/Projects/flutter_spelling_bee/lib/widgets/tally.dart:46:13
When the exception was thrown, this was the stack:
```

It appears that `LinearProgressIndicator` calls `value.round()` which
throws an exception if the value is not finite (`NaN` or `Infinity`).

Modifications:
* We check whether our tally percentage is finite and pass in 0 instead
of `Nan` or `Infinity`.